requirements.txt
numpy
scipy
matplotlib
fpdf
pytest
PyPDF2
pytest-mock
rich
requests
PyYAML
